+++ multimodal/agent-tars-server/src/utils/index.ts
@@ -5,3 +5,4 @@
 
 export * from './workspace';
 export * from './error-handler';
+export * from './workspace-static-server';
